+++ server/api/src/main/java/com/walmartlabs/concord/server/api/project/ProjectEntry.java
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.walmartlabs.concord.common.validation.ConcordKey;
 
 import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Size;
 import java.io.Serializable;
 import java.util.Set;
 
@@ -16,18 +17,27 @@ public class ProjectEntry implements Serializable {
     @ConcordKey
     private final String name;
 
+    @Size(max = 1024)
+    private final String description;
+
     private final Set<String> templates;
 
     @JsonCreator
     public ProjectEntry(@JsonProperty("name") String name,
+                        @JsonProperty("description") String description,
                         @JsonProperty("templates") Set<String> templates) {
 
         this.name = name;
+        this.description = description;
         this.templates = templates;
     }
 
     public String getName() {
         return name;
+    }
+
+    public String getDescription() {
+        return description;
     }
 
     public Set<String> getTemplates() {
@@ -38,6 +48,7 @@ public class ProjectEntry implements Serializable {
     public String toString() {
         return "ProjectEntry{" +
                 "name='" + name + '\'' +
+                ", description='" + description + '\'' +
                 ", templates=" + templates +
                 '}';
     }