+++ server/impl/src/test/java/com/walmartlabs/concord/server/security/SecretDaoTest.java
@@ -22,7 +22,7 @@ public class SecretDaoTest extends AbstractDaoTest {
         String projectName = "project#" + System.currentTimeMillis();
 
         ProjectDao projectDao = new ProjectDao(getConfiguration(), mock(UserPermissionCleaner.class));
-        projectDao.insert(projectName, Collections.emptySet());
+        projectDao.insert(projectName, "test", Collections.emptySet());
 
         String secretName = "secret#" + System.currentTimeMillis();
         SecretDao secretDao = new SecretDao(getConfiguration(), mock(UserPermissionCleaner.class));